requirements.txt
Django==2.2.4
environs==5.2.1
dj-database-url==0.5.0
django-phonenumber_field[phonenumbers]
